test: refactor into separate test files
Test file was becoming too long. Break tests into separate files
grouped by functionality we are testing i.e. the buttons, timer and
redirect/navigation to the page.
